Remove need for mypy ignore directives
Since it can "never" happen that `map_words_in_file` is
called when the spell checker hasn't been initialized,
just assert that to be the case at the start of the routine.

This protects against future code changes accidentally
making it possible (or a bug in the current code).
It also silences mypy.