Don't leave selected range fragments after replace
After replace, or replace all, if there was a selected range,
parts of the range could be left selected, e.g. all the
newlines.
Now, if Replace All in selection only, selection range is
preserved, in case user wants to do something else to
the selected range. Otherwise selection is cleared.