Add Unicode Blocks dialog (#407)
In Tools->Unicode menu.
Also doubles as "Commonly-used characters" dialog, by using the first entry in the dropdown.
It should remember which block was shown the last time the dialog was shown.
Clicking a character should insert it. Hovering over a button should show a tooltip with the character ordinal & name.