Improve efficiency of code
Instead of counting number of newline in slurped text
to get the end point of the slurp, just pass it into
the lower function.

This will particularly improve when reverse searching
with lookarounds/backrefs, which do it by multiple
forward searching, each of which would have potentially
counted all the newlines in the file, instead of just doing
that once.