Add fixup tools for Illustrations and Sidenotes
Adds 'Illustration Fixup' and 'Sidenote
Fixup' to the Tools menu.

The tools share the same module and
functions in that module.. There are
two functions used to move a tag up
and two similar functions to move a
tag down.

The movements in either direction
allowed for a tag conform to the
current Guiguts Manual for GG1.

The tools assume the file being
checked have Illustration and Sidenote
tags that conform to the rules set out
in the Formatting Guidelines. If the file
does not conform then some tags will
not be allowed to move if there is a
risk of incorrect deletions after a tag
is moved.

The placement of a tag after a move
always conforms to the Formatting
Guidelines; that is, it will be separated
from other file contents by a blank
line before and after the tag line(s).