Add soft wrapping
Soft Wrap: None/Char/Word in Prefs dialog.

Note that column select doesn't necessarily work well
if you're trying to select stuff in the wrapped part.
That's no surprise, given how column select works,
and that column numbers won't line up if word-wrapped.
I think it would need to be documented, or the whole
idea abandoned - it's not obvious how to fix it.