Update Unicode block list
1. Block list now contains all current Unicode blocks.
2. Dropdown menu still just contains the GG1 list of
blocks (i.e. those that are likely useful for PPing). This
list can easily be updated to include/remove blocks.
3. This means all characters found with the Unicode
Search dialog should now have a block name next
to them, which can be clicked.
4. Note that (3) above means you can show a block
in the Block dialog that isn't in the dropdown. However,
that doesn't seem to be a problem.
5. Especially if you do what is described in 4, you may
get characters showing in the block dialog that are
"new" (i.e post 2002). The tooltip over a "new" character
in the Block dialog will have a warning icon, just like
its description in the Search dialog.

Fixes #421