Try out split screen
I know this has wrong font, wrong, colors, no line numbers
etc. This is just an experiment to see if it would be worth
pursuing. Work so far is minimal - probably quite a bit
more to get it to look smart and check what the bugs or
limitations are, if any.
You can drag the split up & down, but obviously, it also
needs an on/off option somewhere.
Splitting vertically would be possible instead, with a little
more complication if you want the option of either a
horizontal or vertical split.