Update illo_sn_fixup.py
Very significant reduction in code by
reducing the movement of Illo and
SN records to a simple logic.

Abstraction of (repeated) code into functions clarifies processing logic
and shortens the two main functions; the first for movement up and
the second for movement down.

No longer requires Illo or SN record
positioning to conform to DP
Guidelines before it can be moved
since reorganisation of tags may mean
movement temporarily to a position
that is not in conformance; e.g. moving
an Illo or SN tag to the blank line
below a mid-paragraph tag.

Works consistently whether Page
Marker records are present or not.