Fix Illo & SN bugs
Fixed two bugs identified by @srjfoo.
1. A spurious return-character was prefixing the text of most tags in the
dialog;
2. Between refreshing the dialog and
re-highlighting the selected tag after
a move, the highlighting would jump
to the first tag in the dialog before
settling on the selected tag.

Note the fix for 2. above involves a
change to checkers.py.

The labels of the buttons to move
the selected tag up or down now use
Title Case.