Add fixup tools for Illustrations and Sidenotes (#426)
* Add fixup tools for Illustrations and Sidenotes

Adds 'Illustration Fixup' and 'Sidenote
Fixup' to the Tools menu.

The tools share the same module and
functions in that module.. There are
two functions used to move a tag up
and two similar functions to move a
tag down.

The movements in either direction
allowed for a tag conform to the
current Guiguts Manual for GG1.

The tools assume the file being
checked have Illustration and Sidenote
tags that conform to the rules set out
in the Formatting Guidelines. If the file
does not conform then some tags will
not be allowed to move if there is a
risk of incorrect deletions after a tag
is moved.

The placement of a tag after a move
always conforms to the Formatting
Guidelines; that is, it will be separated
from other file contents by a blank
line before and after the tag line(s).

* Update illo_sn_fixup.py

Very significant reduction in code by
reducing the movement of Illo and
SN records to a simple logic.

Abstraction of (repeated) code into functions clarifies processing logic
and shortens the two main functions; the first for movement up and
the second for movement down.

No longer requires Illo or SN record
positioning to conform to DP
Guidelines before it can be moved
since reorganisation of tags may mean
movement temporarily to a position
that is not in conformance; e.g. moving
an Illo or SN tag to the blank line
below a mid-paragraph tag.

Works consistently whether Page
Marker records are present or not.

* Update Illo/SN Fixup

Fixed issues raised by @windymilla
after previous update:

- moving the selected Illustration or
Sidenote record UP/DOWN into a
block of blank lines now moves it to
the top/bottom of the block as with
GG1.

- anomalous Illustration and Sidenote
records such as the mid-paragraph
Sidenote at line 7816 are now flagged
as 'MIDPARAGRAPH' too. This fix has
identified and flagged a second
anomalous mid-paragraph tag at line
8721.

- moving anomalous Illustration and
Sidenote records closes up the line(s) they were moved from.

- The new 'Illustration Fixup' and
'Sidenote Fixup' entries in the Tools
menu each have a '~' keyboard
shortcut.

- The keyboard shortcut for the
Levenshtein 'Edit Distance Check'
too has changed.

* Add fixup tools for Illustrations and Sidenotes

Adds 'Illustration Fixup' and 'Sidenote
Fixup' to the Tools menu.

The tools share the same module and
functions in that module.. There are
two functions used to move a tag up
and two similar functions to move a
tag down.

The movements in either direction
allowed for a tag conform to the
current Guiguts Manual for GG1.

The tools assume the file being
checked have Illustration and Sidenote
tags that conform to the rules set out
in the Formatting Guidelines. If the file
does not conform then some tags will
not be allowed to move if there is a
risk of incorrect deletions after a tag
is moved.

The placement of a tag after a move
always conforms to the Formatting
Guidelines; that is, it will be separated
from other file contents by a blank
line before and after the tag line(s).

* Update illo_sn_fixup.py

Very significant reduction in code by
reducing the movement of Illo and
SN records to a simple logic.

Abstraction of (repeated) code into functions clarifies processing logic
and shortens the two main functions; the first for movement up and
the second for movement down.

No longer requires Illo or SN record
positioning to conform to DP
Guidelines before it can be moved
since reorganisation of tags may mean
movement temporarily to a position
that is not in conformance; e.g. moving
an Illo or SN tag to the blank line
below a mid-paragraph tag.

Works consistently whether Page
Marker records are present or not.

* Update Illo/SN Fixup

Fixed issues raised by @windymilla
after previous update:

- moving the selected Illustration or
Sidenote record UP/DOWN into a
block of blank lines now moves it to
the top/bottom of the block as with
GG1.

- anomalous Illustration and Sidenote
records such as the mid-paragraph
Sidenote at line 7816 are now flagged
as 'MIDPARAGRAPH' too. This fix has
identified and flagged a second
anomalous mid-paragraph tag at line
8721.

- moving anomalous Illustration and
Sidenote records closes up the line(s) they were moved from.

- The new 'Illustration Fixup' and
'Sidenote Fixup' entries in the Tools
menu each have a '~' keyboard
shortcut.

- The keyboard shortcut for the
Levenshtein 'Edit Distance Check'
too has changed.

* Fix Illo & SN bugs

Fixed two bugs identified by @srjfoo.
1. A spurious return-character was prefixing the text of most tags in the
dialog;
2. Between refreshing the dialog and
re-highlighting the selected tag after
a move, the highlighting would jump
to the first tag in the dialog before
settling on the selected tag.

Note the fix for 2. above involves a
change to checkers.py.

The labels of the buttons to move
the selected tag up or down now use
Title Case.

* Group inserts/deletes during a move

All inserts/deletes made during a
move of the selected tag are grouped
so that they can be undone in one go
with Ctrl+Z.

Made an unrelated minor edit to the
docstring of the display_entries
function in checkers.py