Correctly locate spelling errors with single quote
If spelling error was preceded by single straight quote,
code did a second check to see if the word was in the
dictionary without the quote (to cope with the case that
the quote might be an apostrophe `'tis` or an open
quote from speech, etc. However, it didn't correct its
notion of where the word was on the line to account for
the (now missing) quote. It therefore highlighted the line
one space too far left.

Also, noticed during testing & fixed, if the last line of
file didn't have a terminating newline, spell check didn't
spell check that last line.

Fixes #477