Two WF fixes (#486)
* Don't jump in main text with WF letter shortcut

Just show the first word that starts with the given letter,
but don't move the main text to show the location. That
happens if the user clicks on a WF word.

* Make WF count & search consistent

It was counting, but not searching for phrases like
`<i>c</i>4`, i.e. where there was a word character
immediately before/after the marked up phrase.
The `<i>c</i>` should not be found by word count
because it is not a word - the "word" is `c4`.

Fixes #476