Add Stealth Scannos feature
Two scannos files are included with the release:
regex.json and en-common.json, which are based
on the GG1 versions.

The user can also load their own scannos file. The
recently loaded scannos files are in a dropdown at
the top of the dialog.

The first scanno is shown, and the user can click Next
or Prev. If there are no occurrences of the next/prev
scanno, it autoadvances until it finds a scanno that does
occur, or it hits the end/start of the scannos file.

All occurrences of the scanno are listed, and the
replace button (or Cmd/Ctrl left mouse click) will do the
replacement if one is specified in the scannos file.

Scannos can also have an optional hint to explain what
they do, which is displayed in the dialog.