Highlight quote marks & brackets
When activated from the Search menu, this mode finds pairs of quote
marks and brackets surrounding the cursor position and highlights them.
Only pairs are highlighted, and only the nearest surrounding pair (of
each type) surrounding the cursor is highlighted.

Also fixes a failing test (on macOS) in pytest.

Fixes #42