Fix page label bug with Cmd-up/down (#571)
The widget requires both its "selection" and its "focus"
to be set to work properly. Allowing some things to
work on focus and some on selection made it
possible for the selected page to be out-of-sync
with the displayed page when Cmd+up/down was
used in conjunction with plain up/down.