Update footnotes.py
1. Fix review issues.
2. Fix 'Move FNs to Paragraph' bugs.
3. Fix issues identified in tests against
GG1 output. In particular, ensure that
bugs in the GG1 version are not seen
in this GG2 version.
4. Ensures that when footnotes are
moved to LZs or to paragraph end
they are always placed on the correct page.

This is a bug fix of the current
Footnote Fixup and  does not
implement mixed label types
and related changes to re-indexing.

However it provides all the
functionality needed to correctly
process footnotes in all but a few
unusual cases.