Fix column select failures (#589)
* Fix column select failures

1. Column select now works in peer
2. Column select now works if text is scrolled right so that
left-hand edge is off-screen
3. `rowcol` function now uses the focus widget, i.e.
main text or peer. I thought previously it was OK to
just use the main text, but if the index that is being
converted to a rowcol is an `@x,y` string, it matters
which widget you use to convert it. Symptom of
this failing was that if you scrolled the main text
horizontally, when you started a column select in
the peer, the region actually selected didn't match
the region dragged over - it was offset by the amount
that main was scrolled horizontally. The function that
has been changed is used *a lot*, but I'm hoping it
won't have any effect on its other usage.

* Fix column select bug found during review

It was going wrong in the peer - I suspect the reason it
wasn't always easy to reproduce was that it was accessing
maintext instead of the peer at one point, so it either
worked or didn't depending on the exact content of
the relevant part of maintext when the dragging was
occurring in the peer.