Recognize ppgen `.bm` commands
This is to allow a PPer to load a ppgen file that has `.bm`
commands in, and have page image files associated with
the correct place in the file to aid continued work on the
file.

It is not a solution to the issue of ppgen output files
then being loaded by a PPVer using GG2.