Add First/Last button to Search dialog
The First/Last button finds the first occurrence of the string
in the file if searching forwards, and the last occurrence if
searching backwards.

Note that searching backwards can be done by turning on
"Reverse" checkbox. The currently selected value of
"Reverse" is itself reversed if the First/Last button is Shift
clicked, just like the Search button.

So, if "Reverse" is off, click First/Last for the first match,
shift-click it for last match. If "Reverse" is on, click
First/Last for last match and shift-click it for first match.

Since this behavior is identical to the Search button behavior,
it should be intuitive to the user. If not, they don't need to
use the Shift key and can just use the Reverse checkbox.