Fix select-auto-scroll problem on out-of-bounds release (#626)
When select-dragging beyond the widget boundary, releasing the click
would result in the selection converting back to linear, and scrolling
infinitely until clicking back into the window.

This was caused by the event falling back to the default, due to
bind_event() suppressing the original event handler. Change this to an
additional binding to allow the normal release behavior to be processed.

This commit also adds a check to stop calling the scroll handler when
the boolean is False, moves the boolean being set into the click and
release handlers (instead of deeper down where it also affected the
shift-click mechanism), and uses the event widget reference instead of
relying on focus_widget as it was doing before.

This is a followup to #596 and #620.