Use consistent definition of word in WF
Because regexes include underscore in their idea of word
characters, an occurrence preceded by underscore of a
word found by WF would not be found by "whole word"
search. So, the word would be listed in WF, but when the
user clicked it, the word would not be found.

Fixed by changing search's "whole word" to wrap the
search string with a more complex regex, rather than
`\b...\b`

Also changed WF's idea of a word to include embedded
underscores (in the same way as embedded periods are
included), e.g. "i_001.jpg", but to strip leading or trailing
underscores, which signify italics, e.g. "_dog_".

Fixes #628