Improve detection of mid-para illos/SNs
It now looks forward to find the first "normal" line, i.e.
not an empty line, a `[Blank Page]`, another illo/SN, nor
a page separator line. Then it finds a normal line, it checks
if the line above it is blank, meaning it's the start of a
paragraph. If not, then the illo/SN is mid-paragraph.

Fixes #595