Playing around with highlight colors.
The main one I really want to see changed is the spotlight
color, from orange to darkorange, at least for the dark
themes -- plain orange just doesn't show the white as well.
The rest is playing around only for the colors for the dark
themes to try to get better contrast between text color and
background color.