Add default extension during File-->Save As
First detect if the file is an HTML file, i.e. has the first line
`<!DOCTYPE html>`
Also check if the filename currently has an extension.
If no extension, or extension is `.txt`, then suggest `.html`
for HTML files.
If no extension suggest `.txt` for non-HTML files.
Use the suggested extension when pre-filling the
Save As dialog. User can override and set their own
extension if they want to.

Fixes #651