Swap/delete around curly quote if requested
Previously swap/delete space didn't work if there was a
space on both sides of the quote. Now, depending on
if it was an open/close quote, it swaps/deletes the
space after/before the quote respectively.

Also stop focus being accepted by the buttons, so that
you can immediately use Cmd/Ctrl-Z to undo if you
change your mind.

Fixes #667