Add checker dialog undo update mechanism (#669)
If a checker dialog makes a change, e.g. moving an
illustration down to a paragraph break, it may re-run the
tool. However, if the user does Undo or Redo, there was
no way to flag that the tool needed re-running again.

This commit adds such a mechanism. The mechanism
is based within maintext, since that is where Undo/Redo
occur, but convenience functions have been added to
CheckerDialog because it is probably checkers that will
want to take advantage of it.

The checker dialog registers a function with maintext
to be called on undo/redo & removes it when destroyed.

Implemented for Illo/SN fixups, as an example of what
could be done for any checker tool - it isn't needed for
all checkers.

In this commit, if you do an undo/redo, the list of
messages in the illo/SN checkers is replaced with a
prompt to click the message to refresh.