Fix some  HTML autogen bugs (#686)
* Fix couple HTML autogen errors

1. If the final chapter heading is incorrectly formatted,
having 4 blank lines before, but not 2 after, the code
never found the end of the chapter heading, and so
later went wrong looking for `</h2>`, leading to
an exception. Fixed by trapping at end of file if code
thought it was still in a chapter heading. Fixes #684
2. When an error is trapped by raising an exception,
the code then continued to run, when it should
have stopped.
3. The code stopped 1 line from the end, so a final
footnote was reported as not having a blank line
following it, although it did.

* Improve detection of title

Trim trailing/leading spaces from suspected title, which
also leads to capitalization working correctly.

Fixes #685