Add unmatched HTML tag check
1. In the HTML menu, Unmatched HTML tags reports on
open tags that are not closed, or close tags that were
not opened.
2. It does not attempt to check whether the HTML is
valid, for example, putting an `h1` inside a `span`. The
rules on what is valid are too complex, and the rules
on what would make sense syntactically even more so.
Use the HTML validator to check if HTML is valid.

Fixes #454