Re-do "highlight all matches" to on_change redraw
Rather than redraw the "match all" highlights for the entire file
when the "Search" button is clicked, instead keep track of the last
search string (from when Search was clicked) and whether or not we
should be highlighting all matches currently. Redraw in on_change
both for efficiency in large files, and to update the view more
accurately if the file contents are edited.

Highlights should clear as they did previously:
- When search dialog is closed with mouse
- When search dialog is closed with keyboard shortcut
- If "Remove all highlights" is run
- If a new saerch is started, highlights removed first before adding
  highlights for new search
- If search has no results, then no highlight (because they were
  already removed)

Fixes #705