Image viewer & next/prev page improvement (#753)
Two changes which are independent to improve the
interplay between the text window prev/next page and
the image viewer prev/next file.

For discussion as to whether either, neither or both of
these are improvements:
1. When using prev/next image buttons in status bar
(or equivalent buttons in Search-->Goto menu) step
back/forward to the next page boundary, even if that
means the insert cursor doesn't move. This happens if
there are two or more page breaks at the same point
in the file, due to full page illo or blank page.
2. Don't turn off Auto Image when prev/next file buttons
are used in image viewer. Instead, just pause it until the
user moves the mouse out of the image viewer. So the
PPer can easily look back/forward a page or two around
their current insert location, then when they want to
go back to the text to continue editing, auto image will
resume. Of course, if the user doesn't want auto image
they can always turn it off.