Add shortcuts to Spell Check dialog (#756)
For ease of transition, it uses the same shortcuts as the
GG1 Spell Check. Hovering over the buttons will show a
tooltip with the shortcut.

Note - to have keyboard shortcuts requires the dialog
to have focus. On non-Mac systems, clicking on a
message normally transfers focus to the text window.
This has been suppressed for the Spell Check dialog.

Something of an experiment to see if it improves the
experience of using the dialog.