Add buttons to checker dialogs
There are either 2 or 6 buttons, depending on whether
the checker is able to Fix a reported problem. The buttons
should resize with the dialog width.

Each button has a tooltip, and clicking on each one is
equivalent to using Ctrl/Cmd/Shift Left/Right mouse
buttons in the message list, but just using a simple
mouse click.

Fixes #752