Changes from tests on different platforms
If user does not give an app name, then attempt to open
with the default app.

Also improve how it copes when the pngs dir isn't found.
Each time user turns Auto Img on, or clicks See Img,
prompt them for the pngs dir if the scan file isn't found.