Track image auto-fit with change in viewer size
Bug-fix: when auto-fit modes are on, and the height or width of the viewer changes (whichever is relevant), the image was not redrawn until a new image was loaded.

Change of behavior: if user manually zooms (with +/- buttons or zoom in/out hotkeys including cmd/ctrl-zero), autofit modes are disabled.

Fixes #804