Fix macOS clipboard problem more widely
When Tk copies text to the clipboard, it doesn't update
macOS's clipboard counter. Some apps use this counter
to detect when the clipboard has changed, so don't
realize there is new text there.

The fix is to use macOS `pbcopy` command to "touch" the
clipboard by non-Tk means.

This was previously implemented for copy/cut within the
main text window. However, the Copy Results buttons in
checker dialogs and WF, and the Copy Version Info in
Help About Guiguts didn't have the fix.

Code altered to override the default clipboard append
routine for maintext, and also to use maintext's version
for the other cases.

Hopefully this means that anyone in the future looking
how to copy to the clipboard will use the new functions,
and thus circumvent the bug.