Fix manual copying from read-only texts
The most obvious place this will affect is Help About
version numbers, when you manually select & copy
instead of using the button.
The other is if you use Cmd+a to select the whole of
the text in a checker dialog (like Find All), then Cmd+c
to copy it.