Select lines by dragging over line-number gutter
Clicking and releasing on the line number gutter selects the clicked
line. Clicking and dragging in either direction selects the range of
lines.

Caveat: due to the implementation of the line gutter, currently dragging
beyond the visible range of lines is not supported. It may be possible
with further work, but I didn't attempt it.