Add "Gray out unused view options" button
In View Options dialog (from Bookloupe checker) add a
checkbox to gray out any view options that are not
relevant to the current set of messages, i.e. if no messages
of that type are shown in the checker dialog.

Gray-ness should be updated live, i.e. if a user hides the
last message of a given type, and the View Options dialog
is visible, the view option for that type will be grayed out.
And if the user re-runs the tool, causing a message of the
type to appear in the list again, the view option should be
ungrayed.

This is to support a common way of using View Options
as described in #723

Fixes #723