Improve "see"ing matches in the text (#816)
* Improve "see"ing matches in the text

Since the in-built feature to "see" a bit of text only accepts
a single location, it might end up right on the edge of the
screen horizontally (I already fixed the similar vertical
limitation). To improve this, instead of seeing the start
of a search (or checker dialog text position, e.g. in Find
Proofer Comments) "see" the mid-point of the search,
but only if it's not a multi-line match.

Fixes #814 (after a fashion)

Minor improvement

* Catch multi-line comments

Previous commit broke the detection of multi-line
comments.

Also make Highlight Proofer Comments cope with
multiline comments