RTL experiment
Add advanced "Autofix RTL text" preference.
If this is False there is no change in behavior.
Defaults to False on Macs and True on Windows/Linux.

When a file is loaded, sections of Hebrew (ATM) are
word-order reversed, i.e. order of chars within word
is unchanged, just order of words changed.
When a file is saved, the same happens.
Also, and extra "Paste RTL text" menu option added
for pasting in Hebrew from the forums.