Quotes view options (#828)
* Add View Options to quotes

Bookloupe was an example of using View Options to
toggle show/hide based on error message text, whereas
this is an example of basing it on the error prefix.

First version has checkbox for every message type.

* Reduce number of View Options

Just have two view options for double quotes (unconverted
ones and "other" message) and same for single quotes.

It's actually more helpful to have all double quote errors
(apart from unconverted) because otherwise if you look
to see why you got "Unexpected open double" it may
be because of an error earlier in the paragraph, but you
didn't notice because that error was hidden.

So, I think this is the best split of View Options.