Implement active/inactive pane cursorline (#878)
* Improve peer cursor line highlighting

Highlight both the cursor line in maintext and the
cursor line in peer. This is an improvement on just
one of the cursor lines being highlighted.

Fixes #863

* Implement active/inactive pane cursorline

Building on the previous commit, separate CURSOR_LINE highlight
into CURSOR_LINE_ACTIVE and _INACTIVE. The highlights will be
swapped between panes depending on which is currently active,
rather than using "main" vs "peer" which in this project means
essentially "top" vs "bottom" pane.

The "active" highlight is meant to be more prominent than the
"inactive". In the case where the viewport is covering the same
small area in both panes, the highlights will co-exist and their
being different will help to differentiate which is which within
each pane.

The "inactive" highlight is lower priority than the "active" so
if they cover the same line, the active takes precedence. This
means in the pathological case where the cursor is on the same
line in both panes, both panes will appear to have the active
cursor line. There's no workaround for that edge case (so far
as I know).

Fixes #863

---------

Co-authored-by: Nigel <windymilla@users.noreply.github.com>