Add image short-name label to image viewer control strip (#911)
Rather than let the user guess which image is being displayed, show it to them. The path leading up to the file, as well as the file extension, are stripped to create a short label (i.e. "001.png" becomes "001", and "washington.jpg" becomes "washington"). The label width adjusts dynamically to the label length.

When no image is loaded, the text "<no image>" is used as a placeholder.

Fixes #910