Improve Stealth scannos (#943)
* Improve stealth scanno behavior

1. Only output "Check complete" message when it gets
to the last Scanno in the file.
2. Auto-start the scanno checks when the dialog is
popped or a new file loaded.

* Further improvments

3. Add Auto Advance checkbox - necessary to allow
user to switch it off if they want to swap terms,
or to manually adjust the regexes.
4. Make the search/replace fields be editable
5. Add Swap Terms button to swap search/replace
6. Trap bad regexes & report via error rather than
traceback.
7. Display number of scanno regex being displayed,
e.g. "7/23"
8. Add Replace All button to replace all occurrences
shown in list. This deliberately does not remove the
messages in the list, so the user can click through if
they want to check the fix was done correctly.