Output warning message if shortcut reassigned
If the original definition didn't match the new one, the
message wasn't output. An example would be
`Cmd/Ctrl+O` as the default for File->Open, and the
user attempts to assign `Ctrl+Key-O` or `Cmd+Key+O`
to another command.

Solved by processing both the intended new shortcut
and the original shortcut before comparing them. In
the example above, this converts them both to
`Ctrl+O` or `Cmd+O` depending on platform.