Don't mark ppgen files as edited when loaded (#965)
This was mostly an artifact of re-using the page marker
flags method of updating page positions. Not required
and was annoying to users.

Fixes #961

Testing notes: Load a ppgen master file, e.g. the one
from the linked issue. It should not show "edited"
until a further edit is made.