Improve setting of shortcut keystrokes (#962)
* Improve setting of shortcut keystrokes

Allow user to press key with modifiers instead of using
checkboxes to set the modifiers.

Should also allow use of option key on macOS.

* Detect keyboard better

* Fix Set Bookmark shortcuts for Linux

Use Tk names for `!@#$%"` which Tk on Linux
apparently wants, but Windows/macOS don't.

Also, fix macOS keyboard detection. It should now
detect keyboards that begin with "British", "gb", or
"00000809" as UK keyboards, and all others as US.
Still currently only used to determine which default
shortcuts to define for Set Bookmark menu buttons.

* Insist on Ctrl/Cmd/Alt in shortcuts

Should fix some reported behavior particularly relating
to use of unmodified Option key.

* Couple more bug fixes

1. Clear modifier flags if dialog loses focus, otherwise
Alt-tab to switch focus to another window will cause
the Alt modifier flag to be set, but the release of the
Alt key will not be when the dialog has focus.
2. Do a dummy bind of the requested keystrokes to
test if it's legal - if not, show error message, rather
than an exception.

* Don't allow Option for user-defined shortcuts

* Remove apply button

* Output warning message if shortcut reassigned

If the original definition didn't match the new one, the
message wasn't output. An example would be
`Cmd/Ctrl+O` as the default for File->Open, and the
user attempts to assign `Ctrl+Key-O` or `Cmd+Key+O`
to another command.

Solved by processing both the intended new shortcut
and the original shortcut before comparing them. In
the example above, this converts them both to
`Ctrl+O` or `Cmd+O` depending on platform.

* On Linux, destroy dialog instead of re-using it

Some Linux Window Managers appear to stop the
program from raising a dialog to the top of the
stacking order programmatically. Destroy and
recreate the dialog instead.

Fixes #970

Also fix HTML validate not clearing dialog when re-run,
so messages become duplicated.

* Implement less extreme dialog raising

Although this version doesn't work for WSL, no real user
would be using that to run Guiguts.

* Restore correct F key checks

1. F keys are permitted to be used without modifiers.
2. Plain F1 is not permitted as a shortcut.

* Don't let user reassign Cmd/Ctrl+A,C,V,X

* Don't add cruft to command palette entry

Typing keystrokes with focus in command list passed
the base character in the event. Ignore events that
have Ctrl, Cmd or Alt pressed.

* Lint fixes

* Fix bug introduce by previous commit

Pressing a non-printable key caused the dialog list to
be refreshed, and the top item to be selected.

* Allow removal of shortcuts

1. Restore behavior of Delete/Backspace to delete the
current shortcut. Note that Ctrl+Delete, for example
can still be set up as a shortcut, just not plain Del/BS.
2. Also disallow `Cmd+?` being used as a shortcut

* Disallow Cmd+Shift+?

* Allow clearing of shortcut

Empty shortcut should not complain that it doesn't
have Ctrl/Alt/Cmd modifiers.