Add more shortcuts (#973)
* Add more shortcuts

1. Add Quit Without Saving
2. Add to SR dialog: Search, Replace, Replace All, and Replace & Search (also reverse versions for Searches)
3. Permit Ctrl/Cmd/Shift/Alt plus Return key to be used
for shortcuts (not plain Return)

* Fix bug with Return key

1. You couldn't type Return in main text window.
2. Also added, Find First/Last, Find All, and Count
as shadow commands

* Add S/R checkboxes as shadow commands

Note need for calls to command on/off functions for
some checkboxes.

* Remove Ctrl+N/P shortcuts in Cmd Palette

But add "select next/previous" as shadow commands
so user can set up shortcut if desired.

* Only close Cmd Palette with plain return

Not with Ctrl/Cmd/etc as they might be user shortcuts
for other operations.